http turn screen

+++ app/src/main/java/com/eturn/data/Positions.kt
@@ -1,8 +1,8 @@
 package com.eturn.data
 
 data class Positions(
-    var id: Int,
+    var id: Long,
     var name: String,
     var groupNumber:String,
-    var idUser: Int
+    var idUser: Long
 )
