+++ app/src/main/java/com/eturn/data/Turn.kt
@@ -1,10 +1,10 @@
 package com.eturn.data
 
 data class Turn(
-    var id: Int,
+    var id: Long,
     var name: String,
     var description: String,
     var creator: String,
-    var userId : Int,
+    var userId : Long,
     var countUsers: Int
 )
